+++ crates/vhc/host/daemon-vhc-host/tests/trainer_goldens.rs
@@ -1323,9 +1323,10 @@ mod gpu {
         type Bridge = NoBridge;
         type Client = RecClient;
         type FloatElem = f32;
-        type IntElem = i64;
-        // Mirrors the guest SDK channel (`daemon-vhc-sdk-compute`): bool rides the wire as u32
-        // storage — WGSL forbids native bool in the storage address space (cubecl#1274).
+        // Mirrors the guest SDK channel (`daemon-vhc-sdk-compute`): i32 indices (i64 kernels are
+        // DXC-only on DX12) and u32 bool storage — WGSL forbids native bool in the storage
+        // address space (cubecl#1274).
+        type IntElem = i32;
         type BoolElem = u32;
 
         fn name(_device: &NdArrayDevice) -> String {
